backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
httpx==0.25.2
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.25.2
feedparser==6.0.10
lxml==4.9.3
openai==1.3.7
PyPDF2==3.0.1
python-docx==1.1.0

# Email Agent Dependencies - Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv>=1.0.0 